+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added 3D spatialisation settings to Playback.cs.
 - Replaced SoundEffect constructor with static method.
 
 ## [1.0.3] - 2022-02-27